non-trivial cinit?
Former-commit-id: 1ba7c9f140043284cd937da6ff6b0fd5f8befd07

+++ models.pxd
@@ -12,7 +12,7 @@ cdef class Model:
             np.ndarray  __nudges
             dict neighbors
             dict weights
-            long _nStates
+            int _nStates
 
         #private
         dict __dict__
